+++ components/brave_ads/core/internal/account/utility/refill_unblinded_tokens/request_signed_tokens_url_request_builder.cc
@@ -102,7 +102,7 @@ std::string RequestSignedTokensUrlRequestBuilder::BuildSignatureHeaderValue(
   std::string concatenated_header;
   std::string concatenated_message;
 
-  unsigned int index = 0;
+  size_t index = 0;
   for (const auto& [header, value] : headers) {
     if (index != 0) {
       concatenated_header += " ";
